add StickerColors on presentation layer's color palette.
Color can be regarded as a ULong type value not a Int type value.

+++ domain/src/main/java/com/greedy0110/domain/ColorStore.kt
@@ -1,6 +1,6 @@
 package com.greedy0110.domain
 
 interface ColorStore {
-    suspend fun getOrder(): List<Int>
-    suspend fun saveOrder(newOrder: List<Int>)
+    suspend fun getOrder(): List<ULong>
+    suspend fun saveOrder(newOrder: List<ULong>)
 }
